feat: add AWS S3 cloud storage support

+++ src/main/kotlin/domain/model/file/FileDownloadInfo.kt
@@ -3,5 +3,5 @@ package domain.model.file
 data class FileDownloadInfo(
     val downloadUrl: String,
     val fileName: String,
-    val expiresInSeconds: Long
+    val expiresInSeconds: Int
 )
