+++ src/main/kotlin/domain/model/file/FileUploadInfo.kt
@@ -3,5 +3,5 @@ package domain.model.file
 data class FileUploadInfo(
     val uploadUrl: String,
     val fileKey: String,
-    val expiresInSeconds: Long
+    val expiresInSeconds: Int,
 )
